feat(EPIC-001): Phase 2c — sov:: arithmetic module, eliminate 8-decimal constants
Add lib_types::tokenomics::sov module as single source of truth for SOV
decimal conversions. All SOV amount calculations now go through:
  sov::atoms(whole_tokens) — convert to 18-decimal atomic units
  sov::to_display(atoms)   — convert to human-readable f64
  sov::SCALE              — raw 10^18 scale factor

Migrated ALL 8-decimal SOV constants:
- SOV_ATOMIC_UNITS: 100_000_000 → sov::SCALE (4 files)
- Welcome bonus: 5_000 * 100_000_000 → sov::atoms(5_000)
- UBI monthly: sov::atoms(1_000)
- Validator rewards: sov::atoms(100)
- PoW budget: sov::atoms(21_000_000)
- Domain fee: sov::atoms(10)
- Treasury allocations: all through sov::atoms()
- Display conversions: / 100_000_000 → sov::to_display()

Type widening (u64 → u128):
- WalletInfo.staked_balance, pending_rewards
- PersistedWalletData.balance
- backup_recovery balance tracking
- marketplace payment amounts
- PoW reward accumulators

42 files changed. 1757 tests pass.

Refs: #2093

+++ lib-blockchain/src/blockchain/wallets.rs
@@ -290,11 +290,11 @@ impl Blockchain {
         None
     }
 
-    pub fn collect_sov_backfill_entries(&self) -> Vec<([u8; 32], u64, String)> {
+    pub fn collect_sov_backfill_entries(&self) -> Vec<([u8; 32], u128, String)> {
         let sov_token_id = crate::contracts::utils::generate_lib_token_id();
         let token_opt = self.token_contracts.get(&sov_token_id);
 
-        let mut entries: Vec<([u8; 32], u64, String)> = Vec::new();
+        let mut entries: Vec<([u8; 32], u128, String)> = Vec::new();
         for (wallet_id, wallet) in &self.wallet_registry {
             let is_on_chain = self
                 .wallet_blocks
@@ -344,7 +344,7 @@ impl Blockchain {
     /// Wallets with a balance that differs from the canonical bonus (meaning they
     /// received legitimate transfers on top) are left untouched.
     pub fn correct_ubi_savings_misbalances(&mut self) -> usize {
-        const WRONG_BONUS: u64 = 500_000_000_000; // 5 000 SOV in atomic units
+        const WRONG_BONUS: u128 = 500_000_000_000; // 5 000 SOV in old 8-decimal atomic units
 
         let sov_token_id = crate::contracts::utils::generate_lib_token_id();
         let sov_storage_token_id = crate::storage::TokenId(sov_token_id);
@@ -597,7 +597,7 @@ impl Blockchain {
         &mut self,
         wallet_id: &str,
         recipient_identity: &[u8],
-        amount: u64,
+        amount: u128,
     ) -> Hash {
         let utxo_output = crate::transaction::TransactionOutput {
             commitment: crate::types::hash::blake3_hash(
